pyproject.toml
[project]
name = "python-semantic-versioning"
version = "0.1.0"
description = "An open-source tool developed by DataBurst team to manage the versioning and generating changelog for projects."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-semantic-release>=10.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.12.2",
]
